2/6 deleteUser 탈퇴 기능 추가

+++ src/main/java/com/demo/KIDING/controller/FriendsController.java
@@ -1,15 +1,22 @@
 package com.demo.KIDING.controller;
 
+import com.demo.KIDING.domain.User;
 import com.demo.KIDING.dto.FriendInfo;
 import com.demo.KIDING.dto.MyFriendRes;
 import com.demo.KIDING.global.common.BaseException;
 import com.demo.KIDING.global.common.BaseResponse;
+import com.demo.KIDING.global.common.BaseResponseStatus;
+import com.demo.KIDING.global.jwt.JwtProvider;
+import com.demo.KIDING.repository.UserRepository;
 import com.demo.KIDING.service.FriendsService;
+import com.demo.KIDING.service.UserService;
+import io.jsonwebtoken.Claims;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @Slf4j
 @RestController
@@ -17,6 +24,8 @@ import java.util.List;
 public class FriendsController {
 
     private final FriendsService friendsService;
+    private final UserRepository userRepository;
+    private final JwtProvider jwtProvider;
 
     // 친구 프로필 조회
     @GetMapping("/friends/{userId}/{friendId}")
@@ -28,9 +37,22 @@ public class FriendsController {
         }
     }
 
-    @DeleteMapping("/friends/delete/{userId}/{friendId}")
-    public BaseResponse<String> deleteFriend(@PathVariable Long userId, @PathVariable Long friendId) {
+    @DeleteMapping("/friends/delete/{friendId}")
+    public BaseResponse<String> deleteFriend(@RequestHeader("Authorization") String token, @PathVariable Long friendId) {
         try {
+
+            // 토큰 파싱하여 nickname 추출
+            Claims claims = jwtProvider.parseClaims(token);
+            String username = claims.get("nickname", String.class);
+
+            // nickname으로 User 조회
+            Optional<User> optionalUser = userRepository.findByNickname(username);
+            if (optionalUser.isEmpty()) {
+                throw new BaseException(BaseResponseStatus.NO_USER_FOUND);
+            }
+
+            Long userId = optionalUser.get().getId();
+
             return new BaseResponse<>(friendsService.deleteFriend(userId, friendId));
 
         } catch (BaseException e) {
